// backend/gestion-livraison-de-colis-api/src/common/enum/status-colis.enum.ts
export enum StatusColis {
    EN_ATTENTE = "En attente",
    ANOMALIE = "Anomalie",
    A_CHARGE_DANS_LA_CAMION = "A chargé",
    CHARGE_DANS_LA_CAMION = "Chargé",
    RELIQUAT = "Reliquat",
    EN_EXPEDITION = "En Expedition",
    EN_TRANSIT = "En transit",
    EN_COURS_LIVRAISON = "En cours de livraison",
    RETOUR_EXPEDITEUR = "Retour à l'expediteur",
    DECHARGE_DE_LA_CAMION = "Dechargé",
    LIVRE = "Livré",
}